Added context-based Route maps for Controllers
When a user adds routes in controller context, changes are implicitly
made globally. This is dangerous, anti-intuitive, and breaks
encapsulation. My changes support adding routes as per normal
at the global level, but when using them in controllers, uses the
'addMap' function. This function adds routes in contextual chunks so
that routes matched by one controller won't affect the results in a
different controller.

There is one glaring exception to this rule that exists for reasons of
backwards-compatability. When _ANY_ matching routes were declared in the
global Route object using the regular old 'add' function, _NO_ routes at
the controller level will be notified.

Unit tests are also added to demonstrate the expected result.